requirements.txt
###############################################
# Runtime Dependencies (Pinned for stability)
###############################################
discord.py==2.4.0
python-dateutil==2.9.0.post0
pytz==2024.1

###############################################
# Optional / Internal Helpers
# Uncomment if needed by new features
###############################################
# aiohttp==3.10.5       # If async HTTP requests are ever added
# uvloop==0.20.0        # Faster event loop on Linux
# orjson==3.10.7        # High-performance JSON parser

###############################################
# Development Tools (Uncomment locally only)
###############################################
# pytest==8.3.2
# black==24.10.0
# mypy==1.12.0
# ruff==0.7.2            # Fast linting
# isort==5.13.2          # Import sorting
